requirements.txt
celery
django-celery-beat
django-filter
Django==4.2.13
djangorestframework==3.15.1
fake-headers
requests
bs4
lxml
html5lib
python-dotenv